+++ src/sn_comparator.h
@@ -117,7 +117,7 @@ void sn_comparator_swap (sn_comparator_t *c, int con0, int con1);
 int sn_comparator_compare (const sn_comparator_t *c0,
     const sn_comparator_t *c1);
 
-uint32_t sn_comparator_get_hashval (const sn_comparator_t *c);
+uint64_t sn_comparator_get_hashval (const sn_comparator_t *c);
 
 #endif /* SN_COMPARATOR_H */
